+++ draft-js-side-toolbar-plugin/CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Unreleased
+- expose BlockTypeSelect to make the side toolbar more customizable
+
 ### 2.0.2
 - accomodate editor dom reference for both react 0.15 and 0.16