+++ draft-js-side-toolbar-plugin/CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.4
+- bumped find-with-regex
+
 ### 2.0.3
 - Fixed positioning when the editor has a parent node that is positioned (i.e. `relative`, `absolute` or `fixed`).